requirements.txt
timm==0.4.12
numpy
argparse
pyaml
pillow
opencv-python